refactor: extract binary/container resolution into binary_resolver module
Move the SSL-binary resolution helpers out of main.rs into a dedicated
collector/src/binary_resolver.rs module: resolve_binary_path (PATH/symlink/
shebang resolution), binary_embeds_ssl (static-OpenSSL detection), and the
docker:// container path resolution (parse_container_ref,
resolve_container_binary_path, find_ssl_pid_in_tree), along with their unit
tests. Functions are now pub(crate) and imported by main.rs.

main.rs drops from 1440 to ~1170 lines; the resolution logic is now a cohesive,
independently-testable unit. No behavior change — pure code movement.
